requirements.txt

python-telegram-bot[job-queue]==21.11.1
python-dotenv>=1.0
pandas>=2.0
openpyxl>=3.1

